requirements.txt
pip
selenium
pytest
pytest-html
allure-pytest
PyYAML
zmail